+++ changelog.md
@@ -5,6 +5,7 @@ Uiua is not yet stable.
 ## 0.14.0 - 202?-??-??
 This version is not yet released. If you are reading this on the website, then these changes are live here.
 ### Language
+- Add experimental inline macros
 - Remove the previously deprecated experimental `coordinate ⟔` function
 - Remove the previously deprecated experimental `struct` modifier
 - Remove the previously deprecated `setinv` and `setund` modifiers